+++ src/BaseServiceProvider.php
@@ -10,6 +10,8 @@ class BaseServiceProvider extends ServiceProvider
 {
     protected $commands = [
         \Backpack\Base\app\Console\Commands\Install::class,
+        \Backpack\Base\app\Console\Commands\AddSidebarContent::class,
+        \Backpack\Base\app\Console\Commands\AddCustomRouteContent::class,
     ];
 
     /**
@@ -25,6 +27,13 @@ class BaseServiceProvider extends ServiceProvider
      * @var string
      */
     public $routeFilePath = '/routes/backpack/base.php';
+
+    /**
+     * Where custom routes can be written, and will be registered by Backpack.
+     *
+     * @var string
+     */
+    public $customRoutesFilePath = '/routes/backpack/custom.php';
 
     /**
      * Perform post-registration booting of services.
@@ -46,8 +55,15 @@ class BaseServiceProvider extends ServiceProvider
             __DIR__.'/config/backpack/base.php', 'backpack.base'
         );
 
-        $this->registerAdminMiddleware($this->app->router);
+        // add the root disk to filesystem configuration
+        app()->config['filesystems.disks.'.config('backpack.base.root_disk_name')] = [
+            'driver' => 'local',
+            'root'   => base_path(),
+        ];
+
+        $this->registerMiddlewareGroup($this->app->router);
         $this->setupRoutes($this->app->router);
+        $this->setupCustomRoutes($this->app->router);
         $this->publishFiles();
         $this->loadHelpers();
     }
@@ -81,6 +97,21 @@ class BaseServiceProvider extends ServiceProvider
     }
 
     /**
+     * Load custom routes file.
+     *
+     * @param \Illuminate\Routing\Router $router
+     *
+     * @return void
+     */
+    public function setupCustomRoutes(Router $router)
+    {
+        // if the custom routes file is published, register its routes
+        if (file_exists(base_path().$this->customRoutesFilePath)) {
+            $this->loadRoutesFrom(base_path().$this->customRoutesFilePath);
+        }
+    }
+
+    /**
      * Register any package services.
      *
      * @return void
@@ -111,41 +142,83 @@ class BaseServiceProvider extends ServiceProvider
             if (class_exists('Backpack\Generators\GeneratorsServiceProvider')) {
                 $this->app->register('Backpack\Generators\GeneratorsServiceProvider');
             }
+        } else {
+            $this->checkLicenseCodeExists();
         }
 
         // register the artisan commands
         $this->commands($this->commands);
     }
 
-    public function registerAdminMiddleware(Router $router)
+    public function registerMiddlewareGroup(Router $router)
     {
-        Route::aliasMiddleware('admin', \Backpack\Base\app\Http\Middleware\Admin::class);
+        $middleware_key = config('backpack.base.middleware_key');
+        $middleware_class = config('backpack.base.middleware_class');
+
+        if (!is_array($middleware_class)) {
+            $router->pushMiddlewareToGroup($middleware_key, $middleware_class);
+
+            return;
+        }
+
+        foreach ($middleware_class as $middleware_class) {
+            $router->pushMiddlewareToGroup($middleware_key, $middleware_class);
+        }
     }
 
     public function publishFiles()
     {
-        // publish config file
-        $this->publishes([__DIR__.'/config' => config_path()], 'config');
+        $error_views = [__DIR__.'/resources/error_views' => resource_path('views/errors')];
+        $backpack_base_views = [__DIR__.'/resources/views' => resource_path('views/vendor/backpack/base')];
+        $backpack_public_assets = [__DIR__.'/public' => public_path('vendor/backpack')];
+        $backpack_lang_files = [__DIR__.'/resources/lang' => resource_path('lang/vendor/backpack')];
+        $backpack_config_files = [__DIR__.'/config' => config_path()];
 
-        // publish lang files
-        // $this->publishes([__DIR__.'/resources/lang' => resource_path('lang/vendor/backpack')], 'lang');
-
-        // publish views
-        $this->publishes([__DIR__.'/resources/views' => resource_path('views/vendor/backpack/base')], 'views');
-
-        // publish error views
-        $this->publishes([__DIR__.'/resources/error_views' => resource_path('views/errors')], 'errors');
-
-        // publish public Backpack assets
-        $this->publishes([__DIR__.'/public' => public_path('vendor/backpack')], 'public');
+        // sidebar_content view, which is the only view most people need to overwrite
+        $backpack_sidebar_contents_view = [__DIR__.'/resources/views/inc/sidebar_content.blade.php' => resource_path('views/vendor/backpack/base/inc/sidebar_content.blade.php')];
+        $backpack_custom_routes_file = [__DIR__.$this->customRoutesFilePath => base_path($this->customRoutesFilePath)];
 
         // calculate the path from current directory to get the vendor path
         $vendorPath = dirname(__DIR__, 3);
+        $adminlte_assets = [$vendorPath.'/almasaeed2010/adminlte' => public_path('vendor/adminlte')];
+        $gravatar_assets = [$vendorPath.'/creativeorange/gravatar/config' => config_path()];
 
-        // publish public AdminLTE assets
-        $this->publishes([$vendorPath.'/almasaeed2010/adminlte' => public_path('vendor/adminlte')], 'adminlte');
+        // establish the minimum amount of files that need to be published, for Backpack to work; there are the files that will be published by the install command
+        $minimum = array_merge(
+            $error_views,
+            // $backpack_base_views,
+            $backpack_public_assets,
+            // $backpack_lang_files,
+            $backpack_config_files,
+            $backpack_sidebar_contents_view,
+            $backpack_custom_routes_file,
+            $adminlte_assets,
+            $gravatar_assets
+        );
 
-        // publish public Gravatar assets
-        $this->publishes([$vendorPath.'/creativeorange/gravatar/config' => config_path()], 'gravatar');
+        // register all possible publish commands and assign tags to each
+        $this->publishes($backpack_config_files, 'config');
+        $this->publishes($backpack_lang_files, 'lang');
+        $this->publishes($backpack_base_views, 'views');
+        $this->publishes($backpack_sidebar_contents_view, 'sidebar_content');
+        $this->publishes($error_views, 'errors');
+        $this->publishes($backpack_public_assets, 'public');
+        $this->publishes($backpack_custom_routes_file, 'custom_routes');
+        $this->publishes($adminlte_assets, 'adminlte');
+        $this->publishes($gravatar_assets, 'gravatar');
+        $this->publishes($minimum, 'minimum');
+    }
+
+    /**
+     * Check to to see if a license code exists.
+     * If it does not, throw a notification bubble.
+     *
+     * @return void
+     */
+    private function checkLicenseCodeExists()
+    {
+        if (!config('backpack.base.license_code')) {
+            \Alert::add('warning', "<strong>You're using unlicensed software.</strong> Please ask your web developer to <a target='_blank' href='http://backpackforlaravel.com'>purchase a license code</a> to hide this message.");
+        }
     }
 }