+++ src/app/Http/Controllers/Auth/ResetPasswordController.php
@@ -5,6 +5,7 @@ namespace Backpack\Base\app\Http\Controllers\Auth;
 use Backpack\Base\app\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Password;
 
 class ResetPasswordController extends Controller
 {
@@ -30,7 +31,13 @@ class ResetPasswordController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('guest');
+        $guard = backpack_guard_name();
+
+        $this->middleware("guest:$guard");
+
+        if (!backpack_users_have_email()) {
+            abort(501, trans('backpack::base.no_email_column'));
+        }
 
         // where to redirect after password was reset
         $this->redirectTo = property_exists($this, 'redirectTo') ? $this->redirectTo : config('backpack.base.route_prefix', 'admin').'/dashboard';
@@ -57,5 +64,27 @@ class ResetPasswordController extends Controller
         return view('backpack::auth.passwords.reset', $this->data)->with(
             ['token' => $token, 'email' => $request->email]
         );
+    }
+
+    /**
+     * Get the broker to be used during password reset.
+     *
+     * @return \Illuminate\Contracts\Auth\PasswordBroker
+     */
+    public function broker()
+    {
+        $passwords = config('backpack.base.passwords', config('auth.defaults.passwords'));
+
+        return Password::broker($passwords);
+    }
+
+    /**
+     * Get the guard to be used during password reset.
+     *
+     * @return \Illuminate\Contracts\Auth\StatefulGuard
+     */
+    protected function guard()
+    {
+        return backpack_auth();
     }
 }