+++ src/app/Http/Requests/AccountInfoRequest.php
@@ -3,7 +3,6 @@
 namespace Backpack\Base\app\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class AccountInfoRequest extends FormRequest
@@ -16,7 +15,7 @@ class AccountInfoRequest extends FormRequest
     public function authorize()
     {
         // only allow updates if the user is logged in
-        return Auth::check();
+        return backpack_auth()->check();
     }
 
     /**
@@ -26,7 +25,7 @@ class AccountInfoRequest extends FormRequest
      */
     protected function validationData()
     {
-        return $this->only('email', 'name');
+        return $this->only(backpack_authentication_column(), 'name');
     }
 
     /**
@@ -36,12 +35,12 @@ class AccountInfoRequest extends FormRequest
      */
     public function rules()
     {
-        $user = Auth::user();
+        $user = backpack_auth()->user();
 
         return [
-            'email' => [
+            backpack_authentication_column() => [
                 'required',
-                'email',
+                backpack_authentication_column() == 'email' ? 'email' : '',
                 Rule::unique($user->getTable())->ignore($user->getKey()),
             ],
             'name' => 'required',