+++ src/resources/lang/en/base.php
@@ -9,6 +9,7 @@ return [
     */
 
     'registration_closed'    => 'Registration is closed.',
+    'no_email_column'        => 'Users do not have an associated email address.',
     'first_page_you_see'     => 'The first page you see after login',
     'login_status'           => 'Login status',
     'logged_in'              => 'You are logged in!',
@@ -47,4 +48,11 @@ return [
     'account_updated'        => 'Account updated successfully.',
     'unknown_error'          => 'An unknown error has occured. Please try again.',
     'error_saving'           => 'Error while saving. Please try again.',
+
+    'password_reset' => [
+        'line_1' => 'You are receiving this email because we received a password reset request for your account.',
+        'line_2' => 'Click the button below to reset your password:',
+        'button' => 'Reset Password',
+        'notice' => 'If you did not request a password reset, no further action is required.',
+    ],
 ];