+++ src/resources/views/inc/sidebar.blade.php
@@ -1,4 +1,4 @@
-@if (Auth::check())
+@if (backpack_auth()->check())
     <!-- Left side column. contains the sidebar -->
     <aside class="main-sidebar">
       <!-- sidebar: style can be found in sidebar.less -->
@@ -12,8 +12,8 @@
           <!-- ================================================ -->
           <!-- ==== Recommended place for admin menu items ==== -->
           <!-- ================================================ -->
-          <li><a href="{{ backpack_url('dashboard') }}"><i class="fa fa-dashboard"></i> <span>{{ trans('backpack::base.dashboard') }}</span></a></li>
 
+          @include('backpack::inc.sidebar_content')
 
           <!-- ======================================= -->
           {{-- <li class="header">Other menus</li> --}}